minor code quality improvements
